Towards Python wrapper and Python-based tests
`source/bindings/*` contain pybind11 bindings for internal classes.
There, if necessary, constructors may have a lot of arguments. It is
only important that bindings specify default values, to make user
interface simpler.

For potential future extensions, the compiled .so is wrapped with a
Python module, such that we can add also Python-only code to the final
`cubismup3d` Python module.

Added a basis for Python-based unit tests. Now the questions are:
a) How or what values to export to Python?
b) Python is super slow, how to actually test computed values?
c) How to test with mpirun?