Generate a header file with compile-time settings
If CubismUP3D is compiled with some compile-time flags, it must be
included later with exactly the same values of these flags. Otherwise,
undefined behavior occurs (e.g. different `sizeof(SimulationData)`),
which caused me to spend two hours on debugging and shortened my life
span by two days.

CMake now generates a header file `build/include/CubismUP3D_macros.h`
with all necessary macros. This file is included by `Definitions.h`, but
only if `CUP_INTERNAL_COMPILATION` is not set. With this flag, the
generated file is ignored and instead the compilation relies on compiler
flags.

In fact, from CMake's perspective, this flag is redundant -- the
generated file can always be used. However, it is necessary for the
standard Makefile.