No merged archive, just a list of dependencies.
Prepare a list of dependencies as a space-separated file, instead of
trying to make a merged archive. The applications then load this file
and link all the listed dependencies. The merged archive was broken.
To make it work on both Linux and Mac we'd have to do a weird extract ->
merge trick.

Updated CMake from 3.2 to 3.12.