Use relative header paths
Paths for internal header files should be relative to the current file.
Local header files are put on top, to detect missing includes.

The original idea was to use `#include <CubismUP_3D/absolute path>`
instead of `#include "relative path"`, which might be incompatible with
Windows if one day we need to support it...