Refactor CMake files. Minor changes to bindings
Refactored CMake files into the modern format. I am not certain what to
do about -fPIC which is needed for .so files but affects the
performance. For now I created two copies of the CubismUP_3D core, one
with -fPIC, one without. Probably the best thing would be to put
visibility=hidden by default and then export only those functions that
are needed. However, that would be cumbersome.

Regarding bindings, it is not clear if the constructors for
SimulationData are meaningful. Currently the only reliable interface to
CubismUP_3D are cmdline arguments. Better interface would require
splitting SimulationData into e.g. SimulationData and
SimulationParameters.